Test to reduce jobsystem copy overhead (+150% perf)
This is just a very rough sketch.
Most of the overhead comes from `BulletField` being copied for every job (though I can see the VMs in the future becoming annoying).
A slightly-less-but-still-very painful solution would be to send a `NativeReference<BulletField*>` instead of a raw `BulletField`, but that's stil... uh...

The advantage of doing this is that I can run multiple game intances in parallel. Do I really need that that badly?